requirements.txt
streamlit
google-generativeai
python-dotenv
# tree-sitter
# tree-sitter-python
# tree-sitter-languages
networkx
matplotlib